proto: refactor: extract nar_transfer with split relayed and presigned commits and a bulk cached update
