+++ backend/gradient-proto/src/handler/dispatch.rs
@@ -6,7 +6,7 @@
 
 //! Per-connection message dispatch context and all `ClientMessage` handlers.
 
-use std::collections::{BTreeSet, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 
 use gradient_core::ServerState;
@@ -30,217 +30,11 @@ use super::eval_cache::{
     EvalCacheReceiveStore, handle_eval_cache_chunk, handle_eval_cache_pull, handle_eval_cache_push,
     handle_eval_cache_push_done,
 };
-use super::nar::{NarUploadRecord, mark_nar_stored, record_nar_push_metric};
+use super::nar_transfer::{NarReceiveStore, serve_nar_request};
 use super::socket::{
     JOB_OFFER_CHUNK_SIZE, ProtoWriter, push_pending_candidates, send_credentials_for_job,
-    send_error, send_server_msg, serve_nar_request,
+    send_error, send_server_msg,
 };
-
-// ── Per-session inbound NAR receive store (issue #109, resumable #225) ────────
-
-/// Outcome of [`NarReceiveStore::append`].
-pub(super) enum AppendOutcome {
-    /// Chunk was staged.
-    Ok,
-    /// Fatal: the chunk exceeded the session budget or arrived at a
-    /// non-contiguous offset. The path is now poisoned and its partial
-    /// discarded - the caller aborts the job and rejects the eventual
-    /// `NarUploaded` for the same path.
-    Overflow,
-    /// Chunk arrived for a path the session has already poisoned. Drop it.
-    Poisoned,
-}
-
-#[derive(Default)]
-struct PathState {
-    /// Sender's `stream_token`; empty for legacy pushes that skip the header.
-    token: String,
-    /// Bytes staged for this path on this session (resumed prefix + appends).
-    staged: u64,
-}
-
-/// Disk-backed receiver for inbound `NarPush` chunks. Each push is staged to a
-/// `*.partial` file under `<base_path>/nar-partial/<peer_id>/<hash>` so an
-/// interrupted upload can resume from a byte offset (issue #225) and a large
-/// NAR no longer pins RAM. A per-session byte budget plus a poison set preserve
-/// the #109 protection against a rogue worker opening many un-finalized streams
-/// (the budget now bounds staged **disk**, not RAM). Keying by `peer_id` lets a
-/// reconnecting worker resume its own partial without colliding with another
-/// worker pushing the same content-addressed path.
-pub(super) struct NarReceiveStore {
-    store: gradient_storage::PartialStore,
-    peer_id: String,
-    max_bytes: u64,
-    active: HashMap<String, PathState>,
-    poisoned: BTreeSet<String>,
-}
-
-impl NarReceiveStore {
-    pub(super) fn new(
-        root: std::path::PathBuf,
-        peer_id: &str,
-        ttl: std::time::Duration,
-        max_bytes: u64,
-    ) -> anyhow::Result<Self> {
-        Ok(Self {
-            store: gradient_storage::PartialStore::new(root, ttl)?,
-            peer_id: peer_id.to_owned(),
-            max_bytes,
-            active: HashMap::new(),
-            poisoned: BTreeSet::new(),
-        })
-    }
-
-    fn key(&self, hash: &str) -> String {
-        format!("{}/{}", self.peer_id, hash)
-    }
-
-    /// Record the push stream's token and return how many bytes are already
-    /// staged for it (0 on token mismatch / nothing on disk). Clears any stale
-    /// poison so a fresh attempt can proceed.
-    pub(super) async fn note_header(&mut self, store_path: &str, token: &str) -> u64 {
-        self.poisoned.remove(store_path);
-        let received = match store_hash(store_path) {
-            Some(h) => {
-                let (store, key, token) = (self.store.clone(), self.key(h), token.to_owned());
-                tokio::task::spawn_blocking(move || store.received_len(&key, &token).unwrap_or(0))
-                    .await
-                    .unwrap_or(0)
-            }
-            None => 0,
-        };
-        self.active.insert(
-            store_path.to_owned(),
-            PathState {
-                token: token.to_owned(),
-                staged: received,
-            },
-        );
-        received
-    }
-
-    /// Stage a chunk at `offset` (must be contiguous). Creates a token-less
-    /// entry for legacy pushes that skip the header. The blocking disk write
-    /// runs on the blocking pool so the async socket task is never stalled.
-    pub(super) async fn append(
-        &mut self,
-        store_path: &str,
-        offset: u64,
-        data: &[u8],
-    ) -> AppendOutcome {
-        if self.poisoned.contains(store_path) {
-            return AppendOutcome::Poisoned;
-        }
-
-        let Some(hash) = store_hash(store_path) else {
-            return AppendOutcome::Poisoned;
-        };
-
-        self.active.entry(store_path.to_owned()).or_default();
-        let total: u64 = self.active.values().map(|s| s.staged).sum();
-        if total.saturating_add(data.len() as u64) > self.max_bytes {
-            self.poison(store_path, hash).await;
-            return AppendOutcome::Overflow;
-        }
-
-        let token = self.active[store_path].token.clone();
-        let (store, key, data) = (self.store.clone(), self.key(hash), data.to_vec());
-        let len = data.len() as u64;
-        match tokio::task::spawn_blocking(move || store.append(&key, &token, offset, &data)).await {
-            Ok(Ok(())) => {
-                if let Some(s) = self.active.get_mut(store_path) {
-                    s.staged += len;
-                }
-                AppendOutcome::Ok
-            }
-            Ok(Err(e)) => {
-                warn!(%store_path, error = %e, "partial append failed; poisoning path");
-                self.poison(store_path, hash).await;
-                AppendOutcome::Overflow
-            }
-            Err(e) => {
-                warn!(%store_path, error = %e, "partial append task panicked; poisoning path");
-                self.poison(store_path, hash).await;
-                AppendOutcome::Overflow
-            }
-        }
-    }
-
-    async fn poison(&mut self, store_path: &str, hash: &str) {
-        let (store, key) = (self.store.clone(), self.key(hash));
-        let _ = tokio::task::spawn_blocking(move || store.discard(&key)).await;
-        self.active.remove(store_path);
-        self.poisoned.insert(store_path.to_owned());
-    }
-
-    /// A push stream is open for this path (direct mode). `false` means the
-    /// worker uploaded via presigned S3 and there is nothing to commit.
-    pub(super) fn is_active(&self, store_path: &str) -> bool {
-        self.active.contains_key(store_path)
-    }
-
-    /// Actual on-disk length of the staged partial (validating the token).
-    pub(super) async fn committed_len(&self, store_path: &str) -> u64 {
-        let Some(hash) = store_hash(store_path) else {
-            return 0;
-        };
-        let token = self
-            .active
-            .get(store_path)
-            .map(|s| s.token.clone())
-            .unwrap_or_default();
-        let (store, key) = (self.store.clone(), self.key(hash));
-        tokio::task::spawn_blocking(move || store.received_len(&key, &token).unwrap_or(0))
-            .await
-            .unwrap_or(0)
-    }
-
-    /// Read the staged bytes so the caller can commit them to `nar_storage`.
-    pub(super) async fn read_staged(&self, store_path: &str) -> anyhow::Result<Vec<u8>> {
-        let hash = store_hash(store_path)
-            .ok_or_else(|| anyhow::anyhow!("malformed store path {store_path}"))?;
-        let (store, key) = (self.store.clone(), self.key(hash));
-        tokio::task::spawn_blocking(move || store.read_all(&key))
-            .await
-            .map_err(|e| anyhow::anyhow!("read staged NAR task panicked: {e}"))?
-    }
-
-    /// Drop the staged partial and per-path state after a successful commit.
-    pub(super) async fn finish(&mut self, store_path: &str) {
-        self.active.remove(store_path);
-        if let Some(hash) = store_hash(store_path) {
-            let (store, key) = (self.store.clone(), self.key(hash));
-            let _ = tokio::task::spawn_blocking(move || store.discard(&key)).await;
-        }
-    }
-
-    /// Has this path been poisoned by a prior overflow on the same session?
-    pub(super) fn is_poisoned(&self, store_path: &str) -> bool {
-        self.poisoned.contains(store_path)
-    }
-
-    /// Forget the poison flag and discard any partial for `store_path` so a
-    /// later, well-formed retry of the same path can proceed.
-    pub(super) async fn clear_poison(&mut self, store_path: &str) {
-        self.poisoned.remove(store_path);
-        self.finish(store_path).await;
-    }
-
-    pub(super) fn max_bytes(&self) -> u64 {
-        self.max_bytes
-    }
-}
-
-/// Extract and validate the 32-char store-hash from a `/nix/store/<hash>-name`
-/// path. Returns `None` for anything malformed.
-fn store_hash(store_path: &str) -> Option<&str> {
-    let hash = store_path
-        .strip_prefix("/nix/store/")
-        .unwrap_or(store_path)
-        .split('-')
-        .next()?;
-    (hash.len() == 32 && hash.bytes().all(|b| b.is_ascii_alphanumeric())).then_some(hash)
-}
 
 // ── Dispatch context ──────────────────────────────────────────────────────────
 
@@ -490,8 +284,13 @@ impl<'a> DispatchContext<'a> {
     ) {
         let rpc = self.rpc();
         tokio::spawn(async move {
-            rpc.on_worker_metrics(cpu_usage_pct, ram_free_mb, disk_speed_mbps, network_speed_mbps)
-                .await;
+            rpc.on_worker_metrics(
+                cpu_usage_pct,
+                ram_free_mb,
+                disk_speed_mbps,
+                network_speed_mbps,
+            )
+            .await;
         });
     }
 
@@ -947,241 +746,6 @@ impl<'a> DispatchContext<'a> {
             }
         });
     }
-
-    /// Open (or resume) a push stream and tell the worker how many compressed
-    /// bytes are already staged so it can seek its regenerated zstd stream.
-    async fn on_push_stream_header(
-        &mut self,
-        job_id: String,
-        store_path: String,
-        _total_bytes: Option<u64>,
-        stream_token: String,
-        nar: &mut NarReceiveStore,
-    ) {
-        let received = nar.note_header(&store_path, &stream_token).await;
-        debug!(peer_id = %self.peer_id, %job_id, %store_path, received, "NarStreamHeader (push)");
-        let _ = send_server_msg(
-            self.writer,
-            &ServerMessage::NarPushResume {
-                job_id,
-                store_path,
-                received_bytes: received,
-            },
-        )
-        .await;
-    }
-
-    async fn on_nar_push(
-        &mut self,
-        job_id: String,
-        store_path: String,
-        data: Vec<u8>,
-        offset: u64,
-        is_final: bool,
-        nar: &mut NarReceiveStore,
-    ) {
-        debug!(peer_id = %self.peer_id, %job_id, %store_path, offset, is_final, bytes = data.len(), "NarPush");
-        if data.is_empty() {
-            return;
-        }
-        match nar.append(&store_path, offset, &data).await {
-            AppendOutcome::Ok => {}
-            AppendOutcome::Overflow => {
-                let reason = format!(
-                    "NAR upload for {store_path} rejected: staged-partial budget ({} bytes) \
-                     exceeded or non-contiguous offset {offset}",
-                    nar.max_bytes(),
-                );
-                warn!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "poisoning NAR path");
-                self.abort_job(&job_id, reason).await;
-            }
-            AppendOutcome::Poisoned => {
-                debug!(peer_id = %self.peer_id, %job_id, %store_path, "discarding NarPush chunk for poisoned path");
-            }
-        }
-        // The partial is held until `on_nar_uploaded` arrives; that handler
-        // commits it to `nar_storage` and records the metadata atomically so
-        // we never end up with a `cached_path` row claiming bytes that
-        // aren't actually stored.
-    }
-
-    /// Apply the worker's NAR upload metadata.
-    ///
-    /// For direct-mode pushes (preceded by a `NarStreamHeader` + `NarPush`
-    /// chunks), the staged `*.partial` is validated against the reported
-    /// `file_size`, written to `nar_storage`, and only then is
-    /// `mark_nar_stored` invoked. Any failure aborts the job with
-    /// [`ServerMessage::AbortJob`] so the build is marked failed and the
-    /// scheduler does not advertise the path as cached.
-    ///
-    /// For S3 / presigned uploads (no preceding push stream), the worker has
-    /// already PUT the bytes directly to object storage, so the object is
-    /// HEADed and its size compared against the reported `file_size` before
-    /// any metadata is recorded. Skipping that check would let a failed or
-    /// truncated PUT create a `cached_path` row pointing at a missing or
-    /// corrupt object - the zombie class the demote/reconcile machinery
-    /// exists to repair.
-    #[allow(clippy::too_many_arguments)] // mirrors the wire-protocol message fields
-    async fn on_nar_uploaded(
-        &mut self,
-        job_id: String,
-        store_path: String,
-        file_hash: String,
-        file_size: u64,
-        nar_size: u64,
-        nar_hash: String,
-        references: Vec<String>,
-        deriver: Option<String>,
-        nar: &mut NarReceiveStore,
-    ) {
-        debug!(peer_id = %self.peer_id, %job_id, %store_path, %file_hash, file_size, nar_size, %nar_hash, ?deriver, "NarUploaded");
-
-        // Reject any NarUploaded for a path whose chunked transfer was rejected
-        // mid-stream. Without this guard `mark_nar_stored` would record a
-        // `cached_path` row whose bytes never reached `nar_storage` - leaving
-        // the path "cached" in the DB and undeliverable on the next download.
-        if nar.is_poisoned(&store_path) {
-            nar.clear_poison(&store_path).await;
-            let reason = format!(
-                "NarUploaded for {store_path} rejected: prior NarPush chunk \
-                 exceeded the staged-partial budget or arrived out of order"
-            );
-            warn!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "rejecting NarUploaded for poisoned path");
-            self.abort_job(&job_id, reason).await;
-            return;
-        }
-
-        let Some(hash) = store_hash(&store_path) else {
-            let reason = format!("NarUploaded for malformed store path {store_path}");
-            error!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "NarUploaded for malformed store path");
-            self.abort_job(&job_id, reason).await;
-            return;
-        };
-
-        if nar.is_active(&store_path) {
-            let staged = nar.committed_len(&store_path).await;
-            if staged != file_size {
-                let reason = format!(
-                    "staged NAR size {staged} does not match reported file_size {file_size}"
-                );
-                error!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "NAR upload integrity check failed");
-                self.fail_build_transient(&job_id, reason).await;
-                return;
-            }
-            let buf = match nar.read_staged(&store_path).await {
-                Ok(b) => b,
-                Err(e) => {
-                    let reason = format!("failed to read staged NAR: {e}");
-                    error!(peer_id = %self.peer_id, %job_id, %store_path, error = %e, "read staged NAR failed");
-                    self.fail_build_transient(&job_id, reason).await;
-                    return;
-                }
-            };
-            if let Err(e) = crate::ingest::put_nar_idempotent(
-                &self.state.worker_db,
-                &self.state.nar_storage,
-                hash,
-                &file_hash,
-                buf,
-            )
-            .await
-            {
-                let reason = format!("failed to write NAR to storage: {e}");
-                error!(peer_id = %self.peer_id, %job_id, %store_path, error = %e, "nar_storage.put failed");
-                self.fail_build_transient(&job_id, reason).await;
-                return;
-            }
-            nar.finish(&store_path).await;
-            debug!(peer_id = %self.peer_id, %job_id, %store_path, file_size, "NAR stored");
-        } else {
-            match self.state.nar_storage.head_size(hash).await {
-                Ok(Some(size)) if size == file_size => {}
-                Ok(Some(size)) => {
-                    let reason = format!(
-                        "presigned NAR object size {size} does not match reported file_size {file_size}"
-                    );
-                    error!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "presigned NAR upload integrity check failed");
-                    self.fail_build_transient(&job_id, reason).await;
-                    return;
-                }
-                Ok(None) => {
-                    let reason =
-                        format!("presigned NAR object for {store_path} is missing from storage");
-                    error!(peer_id = %self.peer_id, %job_id, %store_path, %reason, "presigned NAR upload integrity check failed");
-                    self.fail_build_transient(&job_id, reason).await;
-                    return;
-                }
-                Err(e) => {
-                    let reason = format!("failed to head presigned NAR object: {e}");
-                    error!(peer_id = %self.peer_id, %job_id, %store_path, error = %e, "presigned NAR HEAD failed");
-                    self.fail_build_transient(&job_id, reason).await;
-                    return;
-                }
-            }
-        }
-
-        let file_size_i64 = file_size as i64;
-        let nar_record = NarUploadRecord {
-            file_hash: &file_hash,
-            file_size: file_size_i64,
-            nar_size: nar_size as i64,
-            nar_hash: &nar_hash,
-            references: &references,
-            deriver: deriver.as_deref(),
-        };
-        if let Err(e) = mark_nar_stored(
-            self.state,
-            self.scheduler,
-            &job_id,
-            &store_path,
-            &nar_record,
-        )
-        .await
-        {
-            warn!(%store_path, error = %e, "failed to mark NAR as stored");
-        }
-        if let Err(e) =
-            record_nar_push_metric(self.state, self.scheduler, &job_id, file_size_i64).await
-        {
-            debug!(error = %e, "failed to record cache metric for NarUploaded");
-        }
-    }
-
-    /// Send `AbortJob` to the worker. Used when a NAR upload cannot be
-    /// committed safely - the worker stops the job and replies with
-    /// `JobFailed`, which the scheduler turns into a failed build.
-    async fn abort_job(&mut self, job_id: &str, reason: String) {
-        let _ = send_server_msg(
-            self.writer,
-            &ServerMessage::AbortJob {
-                job_id: job_id.to_owned(),
-                reason,
-            },
-        )
-        .await;
-    }
-
-    /// A transient server-side NAR storage failure (staged-read or
-    /// `nar_storage` write). Stop the worker and mark the build
-    /// `FailedTransient` directly so the dispatcher re-queues it - a bare
-    /// `abort_job` would be reported by the worker as a permanent failure and
-    /// never retry. The connection is untouched; only this build fails.
-    async fn fail_build_transient(&mut self, job_id: &str, reason: String) {
-        self.abort_job(job_id, reason.clone()).await;
-        if let Err(e) = self
-            .scheduler
-            .handle_job_failed(
-                self.peer_id,
-                job_id,
-                &reason,
-                gradient_types::proto::BuildFailureKind::Transient,
-                &[],
-            )
-            .await
-        {
-            error!(peer_id = %self.peer_id, %job_id, error = %e, "fail_build_transient: handle_job_failed failed");
-        }
-    }
 }
 
 /// Owned handles for the order-independent request/response RPCs, spawned off
@@ -1459,131 +1023,5 @@ mod prunable_known_derivations_tests {
             prunable_known_derivations(candidates, &[o], &unresolved).is_empty(),
             "an edges_unresolved anchor must be re-walked, not pruned"
         );
-    }
-}
-
-#[cfg(test)]
-mod nar_receive_store_tests {
-    use super::{AppendOutcome, NarReceiveStore};
-    use std::time::Duration;
-    use tempfile::TempDir;
-
-    fn assert_ok(o: AppendOutcome) {
-        assert!(matches!(o, AppendOutcome::Ok), "expected Ok");
-    }
-
-    fn store(max_bytes: u64) -> (TempDir, NarReceiveStore) {
-        let dir = TempDir::new().unwrap();
-        let s = NarReceiveStore::new(
-            dir.path().to_path_buf(),
-            "peer-1",
-            Duration::from_secs(3600),
-            max_bytes,
-        )
-        .unwrap();
-        (dir, s)
-    }
-
-    /// A valid 32-char-hash store path keyed by a single repeated char.
-    fn path(c: char) -> String {
-        format!("/nix/store/{}-name", c.to_string().repeat(32))
-    }
-
-    #[tokio::test]
-    async fn append_below_budget_stages_and_reads_back() {
-        let (_d, mut s) = store(1024);
-        let a = path('a');
-        assert_ok(s.append(&a, 0, &[0u8; 256]).await);
-        assert_ok(s.append(&a, 256, &[1u8; 256]).await);
-        assert_eq!(s.committed_len(&a).await, 512);
-        assert_eq!(s.read_staged(&a).await.unwrap().len(), 512);
-    }
-
-    #[tokio::test]
-    async fn non_contiguous_offset_poisons_path() {
-        let (_d, mut s) = store(1024);
-        let a = path('a');
-        assert_ok(s.append(&a, 0, &[0u8; 100]).await);
-        assert!(matches!(
-            s.append(&a, 999, &[0u8; 10]).await,
-            AppendOutcome::Overflow
-        ));
-        assert!(s.is_poisoned(&a));
-        assert!(matches!(
-            s.append(&a, 0, &[0u8; 10]).await,
-            AppendOutcome::Poisoned
-        ));
-    }
-
-    #[tokio::test]
-    async fn append_overflow_poisons_path() {
-        let (_d, mut s) = store(1024);
-        let a = path('a');
-        assert_ok(s.append(&a, 0, &[0u8; 1000]).await);
-        assert!(matches!(
-            s.append(&a, 1000, &[0u8; 100]).await,
-            AppendOutcome::Overflow
-        ));
-        assert!(s.is_poisoned(&a));
-        assert!(matches!(
-            s.append(&a, 0, &[0u8; 50]).await,
-            AppendOutcome::Poisoned
-        ));
-    }
-
-    #[tokio::test]
-    async fn overflow_across_keys_is_caught() {
-        let (_d, mut s) = store(800);
-        assert_ok(s.append(&path('a'), 0, &[0u8; 400]).await);
-        assert_ok(s.append(&path('b'), 0, &[0u8; 400]).await);
-        assert!(matches!(
-            s.append(&path('c'), 0, &[42u8]).await,
-            AppendOutcome::Overflow
-        ));
-        assert!(s.is_poisoned(&path('c')));
-    }
-
-    #[tokio::test]
-    async fn note_header_reports_resumable_prefix() {
-        let (_d, mut s) = store(10_000);
-        let a = path('a');
-        s.note_header(&a, "tok-v1").await;
-        assert_ok(s.append(&a, 0, b"hello").await);
-        // Simulated reconnect: same token resumes; a different token restarts.
-        assert_eq!(s.note_header(&a, "tok-v1").await, 5);
-        assert_eq!(s.note_header(&a, "tok-v2").await, 0);
-    }
-
-    #[test]
-    fn presigned_mode_has_no_active_stream() {
-        let (_d, s) = store(1024);
-        assert!(
-            !s.is_active(&path('a')),
-            "a path with no header/push must not be treated as direct mode"
-        );
-    }
-
-    #[tokio::test]
-    async fn clear_poison_allows_retry() {
-        let (_d, mut s) = store(100);
-        let a = path('a');
-        assert!(matches!(
-            s.append(&a, 0, &[0u8; 200]).await,
-            AppendOutcome::Overflow
-        ));
-        assert!(s.is_poisoned(&a));
-        s.clear_poison(&a).await;
-        assert!(!s.is_poisoned(&a));
-        assert_ok(s.append(&a, 0, &[0u8; 50]).await);
-    }
-
-    #[tokio::test]
-    async fn finish_discards_staged_partial() {
-        let (_d, mut s) = store(10_000);
-        let a = path('a');
-        assert_ok(s.append(&a, 0, b"hello").await);
-        s.finish(&a).await;
-        assert!(!s.is_active(&a));
-        assert_eq!(s.committed_len(&a).await, 0);
     }
 }
