+++ backend/gradient-proto/src/handler/socket.rs
@@ -9,18 +9,14 @@
 //! Pure framing (the [`ProtoSocket`]/[`ProtoReader`]/[`ProtoWriter`]
 //! abstractions and the underlying constants) moved to
 //! [`crate::session::frame`]. What remains here are the helpers that need
-//! access to [`ServerState`] / [`Scheduler`]: NAR serving from
-//! `nar_storage`, credential delivery, and the cached-path self-heal.
+//! access to [`ServerState`] / [`Scheduler`]: job-offer pushes and credential
+//! delivery. NAR transfer moved to [`super::nar_transfer`].
 
-use std::sync::Arc;
-use std::time::Duration;
-
-use futures::StreamExt;
 use gradient_core::ServerState;
 use gradient_types::ids::OrganizationId;
 use gradient_types::*;
 use sea_orm::EntityTrait;
-use tracing::{debug, error, warn};
+use tracing::{debug, warn};
 
 use crate::messages::ServerMessage;
 use gradient_scheduler::Scheduler;
@@ -50,266 +46,6 @@ pub(super) async fn push_pending_candidates(
         )
         .await;
     }
-}
-
-// ── NAR streaming ─────────────────────────────────────────────────────────────
-
-/// Stream a single requested NAR from `nar_storage` to the worker.
-///
-/// Hardening notes:
-/// - The initial storage open is wrapped in `storage_open_timeout`. A stalled
-///   backend (e.g. S3 hung TCP) used to silently consume the dispatch loop's
-///   600 s waiter ceiling; now it surfaces as a `NarUnavailable` within the
-///   open timeout.
-/// - The chunked send path uses [`ProtoWriter`], which bounds per-chunk send
-///   waits via the queue + `send_chunk_timeout` configured at split time.
-///   A stalled peer is detected as `Err(())` from `send_server_msg` and
-///   triggers a best-effort `NarAbort`.
-/// - The body is read from `object_store`'s streaming API - no full file is
-///   ever held in memory. Chunks are coalesced/split to `NAR_PUSH_CHUNK_SIZE`.
-/// - Per-chunk read from the storage stream is also bounded so a backend that
-///   sends the first byte and then hangs cannot pin the task indefinitely.
-pub(super) async fn serve_nar_request(
-    state: &Arc<ServerState>,
-    writer: &ProtoWriter,
-    job_id: &str,
-    store_path: &str,
-    resume_from: u64,
-    client_token: Option<&str>,
-) -> anyhow::Result<()> {
-    let proto_cfg = &state.config.proto;
-    let storage_open_timeout = Duration::from_secs(proto_cfg.nar_storage_open_timeout_secs);
-    let chunk_read_timeout = Duration::from_secs(proto_cfg.nar_send_chunk_timeout_secs);
-
-    let Some(hash) = store_path
-        .strip_prefix("/nix/store/")
-        .and_then(|s| s.split('-').next())
-    else {
-        let reason = format!("invalid store path: {store_path}");
-        send_nar_unavailable(writer, job_id, store_path, reason.clone()).await;
-        return Err(anyhow::anyhow!(reason));
-    };
-
-    let open = |offset: u64| async move {
-        tokio::time::timeout(
-            storage_open_timeout,
-            state.nar_storage.get_stream_from(hash, offset),
-        )
-        .await
-    };
-
-    let (size, mut stream) = match open(resume_from).await {
-        Ok(Ok(Some((size, s)))) => (size, s),
-        Ok(Ok(None)) => {
-            invalidate_cached_path(state, hash, store_path).await;
-            let reason = format!("NAR not found in cache for {store_path}");
-            send_nar_unavailable(writer, job_id, store_path, reason.clone()).await;
-            return Err(anyhow::anyhow!(reason));
-        }
-        Ok(Err(e)) => {
-            let reason = format!("nar_storage.get_stream({hash}) failed: {e}");
-            error!(%store_path, error = %e, "NAR storage read error");
-            send_nar_unavailable(writer, job_id, store_path, reason.clone()).await;
-            return Err(anyhow::anyhow!(reason));
-        }
-        Err(_) => {
-            let reason = format!(
-                "nar_storage.get_stream({hash}) timed out after {}s",
-                storage_open_timeout.as_secs()
-            );
-            warn!(%store_path, "NAR storage open timed out");
-            send_nar_unavailable(writer, job_id, store_path, reason.clone()).await;
-            return Err(anyhow::anyhow!(reason));
-        }
-    };
-
-    // The stored `.nar.zst` is immutable per hash, so the pull token is just
-    // its size. A worker resuming with a stale token (or claiming more bytes
-    // than exist) restarts from 0; the `NarStreamHeader.total_bytes` lets the
-    // worker truncate its `.partial` accordingly.
-    let server_token = format!("len-{size}");
-    let token_mismatch = client_token.is_some_and(|t| t != server_token);
-    let mut start = resume_from;
-    if resume_from > size || token_mismatch {
-        match open(0).await {
-            Ok(Ok(Some((_s, s)))) => {
-                stream = s;
-                start = 0;
-            }
-            _ => {
-                let reason = format!("failed to reopen {store_path} for fresh transfer");
-                send_nar_unavailable(writer, job_id, store_path, reason.clone()).await;
-                return Err(anyhow::anyhow!(reason));
-            }
-        }
-    }
-
-    send_server_msg(
-        writer,
-        &ServerMessage::NarStreamHeader {
-            job_id: job_id.to_owned(),
-            store_path: store_path.to_owned(),
-            total_bytes: size,
-            stream_token: server_token,
-        },
-    )
-    .await
-    .ok();
-
-    let mut buf: Vec<u8> = Vec::with_capacity(NAR_PUSH_CHUNK_SIZE);
-    let mut offset: u64 = start;
-    let mut total: u64 = 0;
-    let mut chunks_sent: u64 = 0;
-
-    loop {
-        let next = tokio::time::timeout(chunk_read_timeout, stream.next()).await;
-        let item = match next {
-            Ok(Some(x)) => x,
-            Ok(None) => break,
-            Err(_) => {
-                let reason = format!(
-                    "NAR storage read stalled > {}s mid-transfer",
-                    chunk_read_timeout.as_secs()
-                );
-                warn!(%store_path, "NAR storage read stall");
-                let _ = send_server_msg(
-                    writer,
-                    &ServerMessage::NarAbort {
-                        job_id: job_id.to_owned(),
-                        store_path: store_path.to_owned(),
-                        reason: reason.clone(),
-                    },
-                )
-                .await;
-                return Err(anyhow::anyhow!(reason));
-            }
-        };
-        let bytes = match item {
-            Ok(b) => b,
-            Err(e) => {
-                let reason = format!("NAR storage stream error: {e}");
-                error!(%store_path, error = %e, "NAR storage stream error");
-                let _ = send_server_msg(
-                    writer,
-                    &ServerMessage::NarAbort {
-                        job_id: job_id.to_owned(),
-                        store_path: store_path.to_owned(),
-                        reason: reason.clone(),
-                    },
-                )
-                .await;
-                return Err(anyhow::anyhow!(reason));
-            }
-        };
-
-        let mut slice = &bytes[..];
-        while !slice.is_empty() {
-            let want = NAR_PUSH_CHUNK_SIZE - buf.len();
-            let take = slice.len().min(want);
-            buf.extend_from_slice(&slice[..take]);
-            slice = &slice[take..];
-            if buf.len() == NAR_PUSH_CHUNK_SIZE {
-                let chunk = std::mem::replace(&mut buf, Vec::with_capacity(NAR_PUSH_CHUNK_SIZE));
-                let chunk_len = chunk.len() as u64;
-                if send_server_msg(
-                    writer,
-                    &ServerMessage::NarPush {
-                        job_id: job_id.to_owned(),
-                        store_path: store_path.to_owned(),
-                        data: chunk,
-                        offset,
-                        is_final: false,
-                    },
-                )
-                .await
-                .is_err()
-                {
-                    let reason = format!("WebSocket send stalled mid-NarPush at offset {offset}");
-                    let _ = send_server_msg(
-                        writer,
-                        &ServerMessage::NarAbort {
-                            job_id: job_id.to_owned(),
-                            store_path: store_path.to_owned(),
-                            reason: reason.clone(),
-                        },
-                    )
-                    .await;
-                    return Err(anyhow::anyhow!(reason));
-                }
-                offset += chunk_len;
-                total += chunk_len;
-                chunks_sent += 1;
-            }
-        }
-    }
-
-    let final_len = buf.len() as u64;
-    if send_server_msg(
-        writer,
-        &ServerMessage::NarPush {
-            job_id: job_id.to_owned(),
-            store_path: store_path.to_owned(),
-            data: buf,
-            offset,
-            is_final: true,
-        },
-    )
-    .await
-    .is_err()
-    {
-        let reason = format!("WebSocket send stalled on final NarPush at offset {offset}");
-        let _ = send_server_msg(
-            writer,
-            &ServerMessage::NarAbort {
-                job_id: job_id.to_owned(),
-                store_path: store_path.to_owned(),
-                reason: reason.clone(),
-            },
-        )
-        .await;
-        return Err(anyhow::anyhow!(reason));
-    }
-    total += final_len;
-    chunks_sent += 1;
-
-    debug!(%store_path, bytes = total, chunks = chunks_sent, "NarRequest served (streaming)");
-    Ok(())
-}
-
-/// Purge a `cached_path` row whose NAR is no longer in `nar_storage`.
-///
-/// Deletes the stale artifact and clears `derivation_output.is_cached` /
-/// `cached_path` so the next `CacheQuery` stops claiming the path is available -
-/// letting the next build either rebuild from source or pick the path up from a
-/// configured upstream. The derivation graph is untouched.
-async fn invalidate_cached_path(state: &Arc<ServerState>, hash: &str, store_path: &str) {
-    match gradient_db::demote_cached_output(&state.worker_db, &state.nar_storage, hash).await {
-        Ok(_) => warn!(
-            %hash,
-            %store_path,
-            "self-heal: NAR missing from storage; cached_path demoted so the path will be rebuilt"
-        ),
-        Err(e) => {
-            warn!(%hash, %store_path, error = %e, "self-heal: failed to demote cached output")
-        }
-    }
-}
-
-async fn send_nar_unavailable(
-    writer: &ProtoWriter,
-    job_id: &str,
-    store_path: &str,
-    reason: String,
-) {
-    let _ = send_server_msg(
-        writer,
-        &ServerMessage::NarUnavailable {
-            job_id: job_id.to_owned(),
-            store_path: store_path.to_owned(),
-            reason,
-        },
-    )
-    .await;
 }
 
 // ── Credential delivery ───────────────────────────────────────────────────────
@@ -372,114 +108,5 @@ async fn send_ssh_key_credential(
         }
         Ok(None) => warn!(%org_id, "org not found for SSH key lookup"),
         Err(e) => warn!(%org_id, error = %e, "failed to fetch org for SSH key"),
-    }
-}
-
-#[cfg(test)]
-mod serve_nar_tests {
-    use super::*;
-    use crate::messages::decode_server_message;
-    use gradient_test_support::state::test_state;
-    use sea_orm::{DatabaseBackend, MockDatabase};
-    use tokio::sync::mpsc;
-
-    /// Spy writer: records every message the server attempted to send so the
-    /// test can assert exactly which protocol frames were emitted (NarPush,
-    /// NarUnavailable, NarAbort, …).
-    fn spy_writer(timeout: Duration) -> (ProtoWriter, mpsc::Receiver<Vec<u8>>) {
-        let (tx, rx) = mpsc::channel::<Vec<u8>>(64);
-        (
-            ProtoWriter {
-                tx,
-                send_chunk_timeout: timeout,
-                _direction: std::marker::PhantomData,
-            },
-            rx,
-        )
-    }
-
-    fn decode(bytes: &[u8]) -> ServerMessage {
-        decode_server_message(bytes).expect("decode ServerMessage")
-    }
-
-    /// Streamed payload arrives as one or more `NarPush` frames whose
-    /// concatenated `data` matches the original bytes, with the final frame
-    /// flagged `is_final=true`.
-    #[tokio::test]
-    async fn serve_streams_full_payload_in_chunks() {
-        let db = MockDatabase::new(DatabaseBackend::Postgres).into_connection();
-        let state = test_state(db);
-        let mut payload = Vec::with_capacity(9 * 1024 * 1024);
-        for i in 0..(9 * 1024 * 1024 / 4) {
-            payload.extend_from_slice(&(i as u32).to_le_bytes());
-        }
-        let hash = "abcdefghijklmnopqrstuvwxyz012345";
-        state.nar_storage.put(hash, payload.clone()).await.unwrap();
-
-        let (writer, mut rx) = spy_writer(Duration::from_secs(5));
-        let store_path = format!("/nix/store/{hash}-test-pkg");
-        serve_nar_request(&state, &writer, "job-1", &store_path, 0, None)
-            .await
-            .expect("serve must succeed");
-
-        let mut assembled = Vec::with_capacity(payload.len());
-        let mut nar_push_frames = 0u32;
-        let mut saw_header = false;
-        let mut saw_final = false;
-        while let Ok(bytes) = rx.try_recv() {
-            match decode(&bytes) {
-                ServerMessage::NarStreamHeader { total_bytes, .. } => {
-                    saw_header = true;
-                    assert!(!saw_final, "header must precede chunks");
-                    assert_eq!(total_bytes as usize, payload.len());
-                }
-                ServerMessage::NarPush { data, is_final, .. } => {
-                    assembled.extend_from_slice(&data);
-                    if is_final {
-                        saw_final = true;
-                    }
-                    nar_push_frames += 1;
-                }
-                other => panic!("unexpected frame: {}", other.variant_name()),
-            }
-        }
-        assert!(saw_header, "a NarStreamHeader must precede the chunks");
-        assert!(
-            nar_push_frames >= 3,
-            "9 MiB / 4 MiB chunks → at least 3 frames, got {nar_push_frames}"
-        );
-        assert!(saw_final, "the last frame must be is_final=true");
-        assert_eq!(
-            assembled, payload,
-            "concatenated NarPush data must equal source"
-        );
-    }
-
-    /// Missing object → `NarUnavailable` (not `NarAbort`, no NarPush) and an
-    /// `Err` from `serve_nar_request`.
-    #[tokio::test]
-    async fn serve_emits_nar_unavailable_when_missing() {
-        let db = MockDatabase::new(DatabaseBackend::Postgres).into_connection();
-        let state = test_state(db);
-        let (writer, mut rx) = spy_writer(Duration::from_secs(5));
-
-        let res = serve_nar_request(
-            &state,
-            &writer,
-            "job-1",
-            "/nix/store/zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz-missing",
-            0,
-            None,
-        )
-        .await;
-        assert!(res.is_err(), "missing path must surface as Err");
-
-        let bytes = rx.try_recv().expect("expect one frame");
-        let msg = decode(&bytes);
-        assert_eq!(msg.variant_name(), "NarUnavailable");
-        assert!(
-            rx.try_recv().is_err(),
-            "no further frames after NarUnavailable"
-        );
     }
 }
